+++ src/nade-create-main/components/EndPosInput.tsx
