+++ src/nade-create-main/components/MiniLabel.tsx
