+++ src/nade-create-main/components/TechniqueSelector.tsx
