+++ src/nade-create-main/components/ThrownFromInput.tsx
