+++ src/nade-create-main/guidelines/GuideLines.tsx
