+++ src/pages/contact.tsx
@@ -6,7 +6,7 @@ import { CsgnInput } from "../common/inputs/CsgnInput";
 import { CsgnTextArea } from "../common/inputs/CsgnTextArea";
 import { AddConctactDTO } from "../admin-main/data/ContactDTOs";
 import { useTheme } from "../store/SettingsStore/SettingsHooks";
-import { SEO } from "../layout/SEO";
+import { SEO } from "../common/SEO";
 import { Dimensions } from "../constants/Constants";
 import { LayoutBuilder } from "../layout/LayoutBuilder";
 import { HeaderDefault } from "../defaultheader/Header";
