+++ src/pages/index.tsx
@@ -1,7 +1,7 @@
 import { GetStaticProps, NextPage } from "next";
 import React from "react";
 import { FrontPage } from "../frontpage/FrontPage";
-import { SEO } from "../layout/SEO";
+import { SEO } from "../common/SEO";
 import { StatsApi, SiteStats } from "../core/api/StatsApi";
 import { LayoutBuilder } from "../layout/LayoutBuilder";
 import { HeaderDefault } from "../defaultheader/Header";
