+++ src/pages/privacypolicy.tsx
@@ -1,7 +1,7 @@
 import { NextPage } from "next";
 import Link from "next/link";
 import { useTheme } from "../store/SettingsStore/SettingsHooks";
-import { SEO } from "../layout/SEO";
+import { SEO } from "../common/SEO";
 import { Dimensions } from "../constants/Constants";
 import { LayoutBuilder } from "../layout/LayoutBuilder";
 import { HeaderDefault } from "../defaultheader/Header";
